Divided different components to different files.
