+++ src/CountdownClock.js
@@ -1,6 +1,4 @@
-import './App.css';
 import React from 'react';
-// import { render } from 'react-dom';
 
 // A class representing a countdown clock.
 class CountdownClock extends React.Component {
@@ -16,15 +14,16 @@ class CountdownClock extends React.Component {
     this.myInterval = null
 
     this.state = {
+      // try replacing props.minutes/seconds with the values above.
       minutes: props.minutes,
       seconds: props.seconds,
       isPaused: true
     }
   }
 
-  //
+
   componentDidUpdate() {
-    
+
   }
 
   // This method is invoked once, and sets a timer.
@@ -90,53 +89,4 @@ class CountdownClock extends React.Component {
   }
 }
 
-//
-class PomodoroTimer extends React.Component {
-
-  // Pomodoro set : 25:00
-  // break set: 5:00
-
-  //
-  constructor(props) {
-    super(props)
-    this.state = {
-      // Setting initial times
-      minutesForTimer: 0,
-      secondsForTimer: 4,
-      isOnBreak: false,
-      pomodoroSetsCounter: 0
-    }
-
-    // bindind the "this"-term in "applyWhenTimerIsDone" to this (PomodoroTimer)
-    this.applyWhenTimerIsDone = this.applyWhenTimerIsDone.bind(this)
-  }
-
-  //
-  applyWhenTimerIsDone() {
-    this.setState(() => ({
-      isOnBreak: !this.state.isOnBreak,
-      minutesForTimer: this.state.isOnBreak ? 0 : 0,
-      secondsForTimer: this.state.isOnBreak ? 2: 3,
-      pomodoroSetsCounter: this.state.pomodoroSetsCounter + 1,
-    }))
-  }
-
-  // todo
-  render() {
-    return (
-      <div>
-        <h1> pomodoro set counter: {this.state.pomodoroSetsCounter} </h1>
-        <h3> Welcome to PomodoroTimer </h3>
-        <h4>
-        <CountdownClock
-          minutes={this.state.minutesForTimer}
-          seconds={this.state.secondsForTimer}
-          applyWhenDone={()=>{this.applyWhenTimerIsDone()}}
-          />
-        </h4>
-      </div>
-    )
-  }
-}
-
-export default PomodoroTimer;
+export default CountdownClock;
